Add pad management support
