+++ src/pad.c
@@ -1,0 +1,113 @@
+#include "curses.h"
+#include <stdlib.h>
+#include <string.h>
+
+/* functions from resize.c */
+extern void _vc_register_window(WINDOW *win);
+extern void _vc_screen_puts(int y, int x, const char *str, int attr);
+
+static WINDOW *pad_root(WINDOW *pad) {
+    WINDOW *r = pad;
+    while (r->parent && r->parent->is_pad)
+        r = r->parent;
+    return r;
+}
+
+WINDOW *newpad(int nlines, int ncols) {
+    WINDOW *win = calloc(1, sizeof(WINDOW));
+    if (!win)
+        return NULL;
+    win->begy = 0;
+    win->begx = 0;
+    win->maxy = nlines;
+    win->maxx = ncols;
+    win->cury = 0;
+    win->curx = 0;
+    win->parent = NULL;
+    win->keypad_mode = 0;
+    win->scroll = 0;
+    win->delay = -1;
+    win->attr = COLOR_PAIR(0);
+    win->is_pad = 1;
+    win->pad_y = 0;
+    win->pad_x = 0;
+    win->pad_buf = malloc(sizeof(char *) * nlines);
+    win->pad_attr = malloc(sizeof(int *) * nlines);
+    if (!win->pad_buf || !win->pad_attr) {
+        free(win->pad_buf);
+        free(win->pad_attr);
+        free(win);
+        return NULL;
+    }
+    for (int r = 0; r < nlines; ++r) {
+        win->pad_buf[r] = malloc(ncols);
+        win->pad_attr[r] = malloc(sizeof(int) * ncols);
+        if (!win->pad_buf[r] || !win->pad_attr[r]) {
+            for (int i = 0; i <= r; ++i) {
+                if (i < r) {
+                    free(win->pad_buf[i]);
+                    free(win->pad_attr[i]);
+                }
+            }
+            free(win->pad_buf);
+            free(win->pad_attr);
+            free(win);
+            return NULL;
+        }
+        memset(win->pad_buf[r], ' ', ncols);
+        for (int c = 0; c < ncols; ++c)
+            win->pad_attr[r][c] = win->attr;
+    }
+    _vc_register_window(win);
+    return win;
+}
+
+WINDOW *subpad(WINDOW *orig, int nlines, int ncols, int begin_y, int begin_x) {
+    if (!orig || !orig->is_pad)
+        return NULL;
+    WINDOW *win = calloc(1, sizeof(WINDOW));
+    if (!win)
+        return NULL;
+    win->begy = 0;
+    win->begx = 0;
+    win->maxy = nlines;
+    win->maxx = ncols;
+    win->cury = 0;
+    win->curx = 0;
+    win->parent = orig;
+    win->keypad_mode = 0;
+    win->scroll = 0;
+    win->delay = -1;
+    win->attr = COLOR_PAIR(0);
+    win->is_pad = 1;
+    win->pad_y = orig->pad_y + begin_y;
+    win->pad_x = orig->pad_x + begin_x;
+    win->pad_buf = orig->pad_buf;
+    win->pad_attr = orig->pad_attr;
+    _vc_register_window(win);
+    return win;
+}
+
+int prefresh(WINDOW *pad, int pminrow, int pmincol,
+             int sminrow, int smincol, int smaxrow, int smaxcol) {
+    if (!pad || !pad->is_pad)
+        return -1;
+    WINDOW *root = pad_root(pad);
+    int rows = smaxrow - sminrow + 1;
+    int cols = smaxcol - smincol + 1;
+    for (int r = 0; r < rows; ++r) {
+        if (pminrow + r >= pad->maxy)
+            break;
+        for (int c = 0; c < cols; ++c) {
+            if (pmincol + c >= pad->maxx)
+                break;
+            int rr = pad->pad_y + pminrow + r;
+            int cc = pad->pad_x + pmincol + c;
+            char ch = root->pad_buf[rr][cc];
+            int attr = root->pad_attr[rr][cc];
+            char buf[2] = { ch, 0 };
+            _vc_screen_puts(sminrow + r, smincol + c, buf, attr);
+        }
+    }
+    return 0;
+}
